+++ src/contracts/index.js
@@ -1,6 +1,8 @@
-import ethers from 'ethers'
+import ethersjs from 'ethers'
 import registryAbi from './abi/registry.abi.js'
 import { strToBytes32, isEmpty } from '../utils/helpers.js'
+
+const { ethers } = ethersjs
 
 export const createProvider = providerUri => {
   if (isEmpty(providerUri)) throw Error('Empty web3 provider uri')