+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ### Changed
 - change naming convention from `timestamp` to `sigTimestamp`
+- change timestamp name for SSO param to `ssoTimestamp`
 
 ## [0.2.0] - 2020-11-04
 ### Added